excetption

+++ springbootexample/src/main/java/com/amigoscode/person/PersonService.java
@@ -1,10 +1,12 @@
 package com.amigoscode.person;
 
+import com.amigoscode.exception.ResourceNotFound;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class PersonService {
@@ -27,6 +29,10 @@ public class PersonService {
 
     public void deletePerson(int id) {
         // business logic
+        Optional<Person> personOptional = personDAO.selectPersonById(id);
+        if (personOptional.isEmpty()) {
+            throw new ResourceNotFound("person with id id: " + id + " not found");
+        }
         personDAO.deletePerson(id);
     }
 }
